[Silver V] Title: 부분 문자열, Time: 8 ms, Memory: 2272 KB -BaekjoonHub

+++ 백준/Silver/6550. 부분 문자열/부분 문자열.cc
@@ -11,8 +11,8 @@ using namespace std;
 int main(){
     string s,t;
     while(cin>>s>>t){
-        long long j = 0;
-        for(long long i = 0; i < t.size(); i++){
+        long j = 0;
+        for(long i = 0; i < t.size(); i++){
             char tw = t[i];
             char sw = s[j];
             if(tw == sw){
